circuit: Add plonk dependency under default feature
