+++ circuits/Cargo.toml
@@ -12,13 +12,23 @@ exclude = [".github/workflows/dusk-ci.yml", ".gitignore"]
 [dependencies]
 dusk-bytes = "0.1"
 phoenix-core = { version = "0.30", default-features = false, path = "../core" }
-dusk-plonk = { version = "0.19", default-features = false }
+dusk-bls12_381 = { version = "0.13", default-features = false }
 dusk-jubjub = { version = "0.14", default-features = false }
-poseidon-merkle = { version = "0.6", features = ["rkyv-impl", "zk", "size_32"] }
-dusk-poseidon = { version = "0.39", features = ["zk"] }
-jubjub-schnorr = { version = "0.4", features = ["zk"] }
-rand = { version = "0.8", default-features = false, features = ["std_rng"] }
-ff = { version = "0.13", default-features = false }
+poseidon-merkle = { version = "0.6" }
+dusk-poseidon = { version = "0.39" }
+jubjub-schnorr = { version = "0.4" }
+dusk-plonk = { version = "0.19", default-features = false, optional = true }
 
 [dev-dependencies]
 lazy_static = "1.4"
+rand = { version = "0.8", default-features = false, features = ["std_rng"] }
+ff = { version = "0.13", default-features = false }
+
+[features]
+default = ["plonk"]
+plonk = [
+    "dusk-plonk",
+    "poseidon-merkle/zk",
+    "dusk-poseidon/zk",
+    "jubjub-schnorr/zk",
+]
